Refactor Text rendering mechanism, add UiDisplayUtil
