+++ src/main/java/project/taskcrusher/model/shared/DateUtilApache.java
@@ -10,7 +10,6 @@ import com.joestelmach.natty.DateGroup;
 import com.joestelmach.natty.Parser;
 
 import project.taskcrusher.commons.exceptions.IllegalValueException;
-import project.taskcrusher.model.event.Timeslot;
 
 //@@author A0163962X
 /**
@@ -81,83 +80,10 @@ public class DateUtilApache {
         return nattyParser.parse(toParse).get(0);
     }
 
-    public static boolean hasPassed(Date date) {
-        assert date != null;
-        Date rightNow = new Date();
-        if (date.before(rightNow)) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    // @@author A0127737X
-    /**
-     * returns a string representation of the given Date in a user-friendly
-     * format
-     */
-    public static String deadlineAsStringForUi(Date date) {
-        assert date != null;
-        SimpleDateFormat formatter;
-        String prepend = "";
-        if (isToday(date)) {
-            formatter = new SimpleDateFormat(PARSE_PATTERNS[FORMAT_DATE_RELATIVE]);
-            prepend = "Today ";
-        } else if (isThisYear(date)) {
-            formatter = new SimpleDateFormat(PARSE_PATTERNS[FORMAT_THIS_YEAR]);
-        } else {
-            formatter = new SimpleDateFormat(PARSE_PATTERNS[0]);
-        }
-        return prepend + formatter.format(date);
-    }
-
-    private static boolean isThisYear(Date d) {
-        Date now = new Date();
-        SimpleDateFormat yearChecker = new SimpleDateFormat("yyyy");
-        return yearChecker.format(now).equals(yearChecker.format(d));
-    }
-
-    private static boolean isToday(Date d) {
-        Date now = new Date();
-        SimpleDateFormat dateChecker = new SimpleDateFormat("yyyyMMdd");
-        return dateChecker.format(now).equals(dateChecker.format(d));
-    }
-
     public static String dateAsStringForStorage(Date date) {
         assert date != null;
         SimpleDateFormat sdf = new SimpleDateFormat(PARSE_PATTERNS[0]);
         return sdf.format(date);
     }
 
-    public static String timeslotAsStringForUi(Timeslot timeslot) {
-        assert timeslot != null;
-        String endFormat, startFormat, prepend = "";
-        if (isSameDate(timeslot.start, timeslot.end)) {
-            endFormat = PARSE_PATTERNS[FORMAT_DATE_RELATIVE];
-        } else {
-            endFormat = PARSE_PATTERNS[FORMAT_DATE_ABSOLUTE];
-        }
-        if (isToday(timeslot.start)) {
-            startFormat = PARSE_PATTERNS[FORMAT_DATE_RELATIVE];
-            prepend = "Today ";
-        } else if (isThisYear(timeslot.start)) {
-            startFormat = PARSE_PATTERNS[FORMAT_THIS_YEAR];
-            if (isThisYear(timeslot.end)) {
-                endFormat = PARSE_PATTERNS[FORMAT_THIS_YEAR];
-            }
-        } else {
-            startFormat = PARSE_PATTERNS[FORMAT_DATE_ABSOLUTE];
-        }
-
-        SimpleDateFormat sdf = new SimpleDateFormat(startFormat);
-        prepend += sdf.format(timeslot.start) + " to ";
-        sdf.applyPattern(endFormat);
-        prepend += sdf.format(timeslot.end);
-        return prepend;
-    }
-
-    private static boolean isSameDate(Date d1, Date d2) {
-        SimpleDateFormat dateChecker = new SimpleDateFormat("yyyyMMdd");
-        return dateChecker.format(d1).equals(dateChecker.format(d2));
-    }
 }
